+++ src/main/java/de/gematik/demis/nps/api/NotificationController.java
@@ -31,21 +31,15 @@ import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.APPLICATION_XML_VALUE;
 
-import com.google.common.base.Strings;
 import de.gematik.demis.fhirparserlibrary.MessageType;
-import de.gematik.demis.nps.config.TestUserConfiguration;
-import de.gematik.demis.nps.error.ErrorCode;
-import de.gematik.demis.nps.error.NpsServiceException;
 import de.gematik.demis.nps.service.Processor;
 import de.gematik.demis.nps.service.response.FhirConverter;
 import jakarta.validation.constraints.NotBlank;
-import java.util.Objects;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 import lombok.extern.slf4j.Slf4j;
 import org.hl7.fhir.r4.model.Parameters;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -60,24 +54,14 @@ public class NotificationController {
 
   public static final String HEADER_REQUEST_ID = "X-Request-ID";
   public static final String HEADER_SENDER = "x-sender";
-  public static final String HEADER_TEST_USER_RECIPIENT = "x-testuser-recipient";
-  public static final String HEADER_IS_TEST_NOTIFICATION = "x-testuser";
 
   private final Processor processor;
   private final FhirConverter fhirConverter;
-  private final TestUserConfiguration testUserConfiguration;
-  private final boolean isNewTestRoutingEnabled;
 
   @Autowired
-  public NotificationController(
-      final Processor processor,
-      final FhirConverter fhirConverter,
-      final TestUserConfiguration testUserConfiguration,
-      @Value("${feature.flag.test_routing_v2}") final boolean isNewTestRoutingEnabled) {
+  public NotificationController(final Processor processor, final FhirConverter fhirConverter) {
     this.processor = processor;
     this.fhirConverter = fhirConverter;
-    this.testUserConfiguration = testUserConfiguration;
-    this.isNewTestRoutingEnabled = isNewTestRoutingEnabled;
   }
 
   @PostMapping(
@@ -105,21 +89,12 @@ public class NotificationController {
       @RequestHeader(CONTENT_TYPE) final MediaType contentType,
       @RequestHeader(value = HEADER_REQUEST_ID, required = false) final String requestId,
       @CheckForNull @RequestHeader(value = HEADER_SENDER, required = false) final String sender,
-      @RequestHeader(value = HEADER_IS_TEST_NOTIFICATION, required = false)
-          final boolean isTestNotification,
-      @CheckForNull @RequestHeader(value = HEADER_TEST_USER_RECIPIENT, required = false)
-          final String testUserRecipient,
       @RequestHeader(value = AUTHORIZATION, required = false) String authorization,
-      final WebRequest request) {
+      final WebRequest request,
+      @Nonnull final TestUserProps testUserProps,
+      @Nonnull final JwtToken token) {
     if ("text".equalsIgnoreCase(contentType.getType())) {
       log.info("sender= {}, deprecated contentType= {}", sender, contentType);
-    }
-
-    final TestUserProps testUserProps;
-    if (isNewTestRoutingEnabled) {
-      testUserProps = getTestUserProps(isTestNotification, testUserRecipient, sender);
-    } else {
-      testUserProps = getLegacyTestUserProps(isTestNotification, sender, testUserConfiguration);
     }
 
     final Parameters response =
@@ -130,52 +105,8 @@ public class NotificationController {
             sender,
             testUserProps.isTestNotification(),
             testUserProps.testUserRecipient(),
-            authorization);
+            authorization,
+            token.roles());
     return fhirConverter.setResponseContent(ResponseEntity.ok(), response, request);
   }
-
-  @Nonnull
-  private TestUserProps getTestUserProps(
-      final boolean isTestNotification,
-      @CheckForNull String testUserRecipient,
-      @CheckForNull String sender) {
-    if (!isTestNotification) {
-      return new TestUserProps(false, "");
-    }
-
-    testUserRecipient = Strings.nullToEmpty(testUserRecipient);
-    if (testUserRecipient.isBlank()) {
-      throw new NpsServiceException(
-          ErrorCode.INVALID_TEST_CONFIGURATION,
-          "Is test request, but no recipient provided with header");
-    }
-
-    sender = Strings.nullToEmpty(sender);
-    if ("$sender".equals(testUserRecipient) && sender.isBlank()) {
-      throw new NpsServiceException(
-          ErrorCode.INVALID_TEST_CONFIGURATION,
-          "Is test request, recipient is $sender but no sender provided with header");
-    }
-    if ("$sender".equals(testUserRecipient)) {
-      testUserRecipient = sender;
-    }
-
-    return new TestUserProps(isTestNotification, testUserRecipient);
-  }
-
-  @Nonnull
-  private TestUserProps getLegacyTestUserProps(
-      final boolean isTestUser,
-      @CheckForNull final String sender,
-      @Nonnull final TestUserConfiguration testUserConfiguration) {
-    // pre feature flag
-    final boolean isTestNotification = isTestUser || testUserConfiguration.isTestUser(sender);
-    String forwardTo = "";
-    if (isTestNotification) {
-      forwardTo = Objects.requireNonNullElse(testUserConfiguration.getReceiver(sender), "");
-    }
-    return new TestUserProps(isTestNotification, forwardTo);
-  }
-
-  private record TestUserProps(boolean isTestNotification, @Nonnull String testUserRecipient) {}
 }