+++ src/test/java/de/gematik/demis/nps/service/codemapping/CodeMappingServiceConfigurationTest.java
@@ -32,13 +32,13 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 class CodeMappingServiceConfigurationTest {
   @Nested
   @SpringBootTest(classes = CodeMappingServiceConfiguration.class)
   class FeatureFlagOn {
-    @MockBean FutsClient futsClientMock;
+    @MockitoBean FutsClient futsClientMock;
     @Autowired CodeMappingService underTest;
 
     @Test